+++ core/dbt/contracts/graph/unparsed.py
@@ -45,6 +45,11 @@ class UnparsedMacro(UnparsedBaseNode, HasSQL):
 
 
 @dataclass
+class UnparsedGenericTest(UnparsedBaseNode, HasSQL):
+    resource_type: NodeType = field(metadata={'restrict': [NodeType.Macro]})
+
+
+@dataclass
 class UnparsedNode(UnparsedBaseNode, HasSQL):
     name: str
     resource_type: NodeType = field(metadata={'restrict': [